feat : order, ingredient, connect 테이블 수정 및 데이터 추가

+++ src/entity/connect.entity.ts
@@ -14,7 +14,7 @@ import { Ingredient } from './ingredient.entity';
 
 @Entity('Connect')
 export class Connect extends BaseEntity {
-  @PrimaryGeneratedColumn({ type: 'bigint' })
+  @PrimaryGeneratedColumn({ type: 'int' })
   id: number;
 
   @ManyToOne(() => Order, (order) => order.connects, { onDelete: 'CASCADE' })
